Use dynamic imports for homepage client components | Move components to more logical locations
